add timeout on tailscale up command
Set a 5 minute timeout when connecting to tailscale to prevent the job
from hanging for hours (#50). This is way longer than should ever be
necessary, but accounts for intermittent networking issues on either the
actions runner or the Tailscale package or control server.  Users can
always set a shorter timeout on their own action step.

Fixes #50

Signed-off-by: Will Norris <will@tailscale.com>